Moved views to resources.py
